rename: 불필요한 뎁스 제거 및 폴더명 수정 (#84)
