+++ app/(dashboard)/strategies/_hooks/query/use-get-details-information-data.ts
