+++ app/(dashboard)/strategies/_hooks/query/use-get-strategies-data.ts
